refactor tests for loading functions
