+++ tests/test_evaluation_tempo.py
@@ -9,7 +9,6 @@ from __future__ import absolute_import, division, print_function
 
 import math
 import unittest
-from os.path import join as pj
 
 from madmom.evaluation.tempo import *
 from . import ANNOTATIONS_PATH, DETECTIONS_PATH
@@ -23,27 +22,6 @@ DET_STRENGTHS = np.asarray([0.6, 0.4])
 
 
 # test functions
-class TestLoadTempoFunction(unittest.TestCase):
-
-    def test_load_tempo_from_file(self):
-        annotations = load_tempo(pj(ANNOTATIONS_PATH, 'sample.tempo'))
-        self.assertIsInstance(annotations, np.ndarray)
-
-    def test_load_tempo_from_file_handle(self):
-        file_handle = open(pj(ANNOTATIONS_PATH, 'sample.tempo'))
-        annotations = load_tempo(file_handle)
-        self.assertIsInstance(annotations, np.ndarray)
-        file_handle.close()
-
-    def test_load_tempo_annotations(self):
-        annotations = load_tempo(pj(ANNOTATIONS_PATH, 'sample.tempo'))
-        self.assertIsInstance(annotations, np.ndarray)
-        self.assertEqual(annotations.shape, (2, 2))
-        self.assertTrue(np.allclose(annotations, ANNOTATIONS))
-        self.assertTrue(np.allclose(annotations[:, 0], ANN_TEMPI))
-        self.assertTrue(np.allclose(annotations[:, 1], ANN_STRENGTHS))
-
-
 class TestSortTempoFunction(unittest.TestCase):
 
     def test_sort(self):
